requirements.txt
pillow~=9.2.0
loguru~=0.5.3
bunch>=1.0.1
numpy>=1.21.6
torchvision>=0.9.0
opencv-python~=4.6.0.66
tqdm>=4.64.1
pyyaml>=6.0
matplotlib>=3.5.3
imgaug>=0.4.0
pandas~=1.3.5
onnx~=1.13.1
onnxsim~=0.4.17
scikit-learn~=1.0.2